+++ code/packages/python/spice-engine/src/spice_engine/model_cards.py
@@ -307,8 +307,8 @@ _MODEL_CARD_SUPPORTED_PARAMETER_KINDS = (
 )
 _MODEL_CARD_SUPPORTED_PARAMETER_COVERAGE_EXPECTED_SUMMARIES = {
     "D": (12, 18, 5, 3),
-    "NPN": (24, 39, 11, 4),
-    "PNP": (24, 39, 11, 4),
+    "NPN": (25, 41, 12, 4),
+    "PNP": (25, 41, 12, 4),
     "NJF": (5, 11, 5, 3),
     "PJF": (5, 11, 5, 3),
     "NMOS": (18, 25, 6, 3),
@@ -383,6 +383,8 @@ _BJT_PARAMETER_ALIASES: dict[str, str] = {
     "ISC": "ISC",
     "NC": "NC",
     "XTB": "XTB",
+    "BR": "BR",
+    "BETA_R": "BR",
     "NF": "NF",
     "NR": "NR",
     "VJE": "VJE",
@@ -935,6 +937,7 @@ def bjt_from_model_card(
         Isc=p.get("ISC", 0.0),
         Nc=p.get("NC", 2.0),
         Xtb=p.get("XTB", 0.0),
+        beta_r=p.get("BR", 1.0),
     )